Delete old StatefulSets for VMAlertmanagers
When the ConfigMap for a VMAlertmanager is updated, VictoriaMetrics
operator should update the StatefulSet for the VMAlertmanager
to refer to the new ConfigMap.
The operator sometimes fails to do this because the old StatefulSet
gets unstable due to the deletion of the old ConfigMap.
The operator cannot reconcile unstable StatefulSets.
So this PR adds a function to delete the old StatefulSet until
the StatefulSet refers to the proper ConfigMap.

Signed-off-by: morimoto-cybozu <kenji_morimoto@cybozu.co.jp>