Disable hash suffix for VMAlertmanager ConfigMaps to stabilize update
This commit disables the hash suffix functionality for the ConfigMaps
used in VMAlertmanager CRs.
The update of ConfigMap contents is detected by the configmap-reload
sidecar containers instead.
With this commit we can avoild the early deletion of old ConfigMaps
by Argo CD, the instability of the VMAlertmanager StatefulSets, and
the hang-up of VictoriaMetrics operator's reconciliation.

Signed-off-by: morimoto-cybozu <kenji_morimoto@cybozu.co.jp>