Rook/Ceph: disable warning by MANY_OBJECTS_PER_PG
Suppress the folowing warning.

```
health: HEALTH_WARN
        1 pools have many more objects per pg than average
```

The definition of this parameter is as follows.

https://github.com/ceph/ceph/blob/d31130d4516a36c585b1224c193ec720f31c649c/src/common/options/mgr.yaml.in#L237-L247
```
- name: mon_pg_warn_max_object_skew
  type: float
  level: advanced
  desc: max skew few average in objects per pg
  fmt_desc: Raise ``HEALTH_WARN`` if the average RADOS object count per PG
    of any pool is greater than ``mon_pg_warn_max_object_skew`` times
    the average RADOS object count per PG of all pools. Zero or a non-positive
    number disables this. Note that this option applies to ``ceph-mgr`` daemons.
  default: 10
  services:
  - mgr
```

This warning is harmless in Neco's Ceph clusters because we don't care the difference of the number of objects per pgs between pools. For example, we hit this problem because there is a big pool and there are many almost empty pools.

Signed-off-by: Yuji Ito <llamerada.jp@gmail.com>